+++ src/Components/TokenPicker/TokenPicker.tsx
@@ -4,7 +4,7 @@ import { FiSearch } from 'react-icons/fi';
 import { useTranslation } from 'react-i18next';
 
 import Input from 'old-components/Guilds/common/Form/Input';
-import { Modal } from 'old-components/Guilds/common/Modal';
+import { Modal } from 'Components';
 import { useAllERC20Balances } from 'hooks/Guilds/ether-swr/erc20/useAllERC20Balances';
 import useMiniSearch from 'hooks/useMiniSearch';
