+++ src/Components/index.ts
@@ -4,3 +4,4 @@ export { NetworkButton } from './NetworkButton';
 export { TokenPicker } from './TokenPicker';
 export { Filter } from './Filter';
 export * from './Web3Modals';
+export { Modal } from './Primitives/Modal';
